requirements.txt
# Core GUI
PySide6==6.6.1

# Core ML stack
tensorflow-cpu==2.13.0
numpy<2.0
scikit-learn==1.3.2

# Image Processing
opencv-python==4.9.0.80

# Utility
pynput==1.7.6